+++ Sources/SwiftCardanoWallet/Models/WalletBalance.swift
@@ -8,15 +8,15 @@ import SwiftCardanoCore
 /// reported by `ChainContext.stakeAddressInfo` — `0` if the stake address isn't registered.
 /// `utxoCount` is the number of UTxOs that contributed to the totals.
 public struct WalletBalance: Sendable, Equatable {
-    public let lovelace: Int
+    public let lovelace: Int64
     public let multiAsset: MultiAsset
-    public let rewards: Int
+    public let rewards: Int64
     public let utxoCount: Int
 
     public init(
-        lovelace: Int,
+        lovelace: Int64,
         multiAsset: MultiAsset = MultiAsset([:]),
-        rewards: Int = 0,
+        rewards: Int64 = 0,
         utxoCount: Int
     ) {
         self.lovelace = lovelace
